user_service/requirements.txt
fastapi[standard]
uvicorn[standard]
sqlmodel
psycopg2-binary
python-dotenv
bcrypt
python-jose[cryptography]
